% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lqa_admm.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Soft Thresholding Operator}
\usage{
soft_threshold(x, lambda)
}
\arguments{
\item{x}{Input vector}

\item{lambda}{Threshold parameter}
}
\value{
Soft-thresholded vector
}
\description{
\eqn{S_\lambda(x) = \text{sign}(x) \max(|x| - \lambda, 0)}
}
\keyword{internal}
