% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lqa_admm.R
\name{solve_dynamic_lqa_admm}
\alias{solve_dynamic_lqa_admm}
\title{Solve Dynamic LQA System using ADMM}
\usage{
solve_dynamic_lqa_admm(
  y_list,
  X_list,
  w_mat,
  lambda,
  tau,
  rho = 1,
  max_iter = 100,
  tol = 1e-06
)
}
\arguments{
\item{y_list}{List of response vectors}

\item{X_list}{List of design matrices}

\item{w_mat}{Weight matrix of dimension \eqn{p \times T} for LQA weights}

\item{lambda}{SCAD tuning parameter}

\item{tau}{Fused penalty tuning parameter}

\item{rho}{ADMM penalty parameter}

\item{max_iter}{Maximum ADMM iterations}

\item{tol}{Convergence tolerance}
}
\value{
Coefficient matrix \eqn{p \times T}
}
\description{
Solves the block tridiagonal system arising from Dynamic SCAD with fused penalty
using Alternating Direction Method of Multipliers (ADMM).
}
\section{Problem Formulation}{


After LQA approximation, we need to solve:
\deqn{
\min_{\boldsymbol{\beta}} \left\{
\sum_{t=1}^T \|\mathbf{y}_t - \mathbf{X}_t \boldsymbol{\beta}_t\|_2^2
+ \sum_{t=1}^T \sum_{j=1}^p w_{j,t} \beta_{j,t}^2
+ \tau \sum_{t=2}^T \sum_{j=1}^p |\beta_{j,t} - \beta_{j,t-1}|
\right\}
}

Using ADMM, we introduce auxiliary variables \eqn{\mathbf{z}_t} and solve:
\deqn{
\min_{\boldsymbol{\beta}, \mathbf{z}} \left\{
\sum_{t=1}^T \|\mathbf{y}_t - \mathbf{X}_t \boldsymbol{\beta}_t\|_2^2
+ \sum_{t=1}^T \sum_{j=1}^p w_{j,t} \beta_{j,t}^2
+ \tau \sum_{t=2}^T \sum_{j=1}^p |z_{j,t}|
\right\}
}
subject to \eqn{\beta_{j,t} - \beta_{j,t-1} = z_{j,t}} for \eqn{t = 2, \ldots, T}.
}

\keyword{internal}
