% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lqa_admm.R
\name{solve_tridiagonal_system}
\alias{solve_tridiagonal_system}
\title{Solve Tridiagonal System using Thomas Algorithm}
\usage{
solve_tridiagonal_system(diag_main, diag_sub, diag_super, rhs)
}
\arguments{
\item{diag_main}{Main diagonal (length n)}

\item{diag_sub}{Subdiagonal (length n-1)}

\item{diag_super}{Superdiagonal (length n-1)}

\item{rhs}{Right-hand side vector (length n)}
}
\value{
Solution vector \eqn{\mathbf{x}}
}
\description{
Solves \eqn{A \mathbf{x} = \mathbf{b}} where A is tridiagonal.
}
\keyword{internal}
